refactor : RedisUtils 작성 및 적용

+++ cache/src/main/java/com/sinor/cache/admin/api/service/IApiServiceV1.java
@@ -30,7 +30,7 @@ public interface IApiServiceV1 {
 	 * @param value 생성할 캐시의 Value
 	 * @param expiredTime 생성할 캐시의 만료시간
 	 */
-	ApiGetResponse saveOrUpdate(String key, String value, int expiredTime) throws CustomException;
+	ApiGetResponse saveOrUpdate(String key, String value, Long expiredTime) throws CustomException;
 
 	/**
 	 * 캐시 삭제
